+++ package.json
@@ -32,6 +32,7 @@
     "ws": "^7.4.3"
   },
   "devDependencies": {
+    "@babel/eslint-parser": "^7.15.7",
     "eslint": "^7.25.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.22.1"
