+++ src/types.js
@@ -1,0 +1,5 @@
+/**
+ * @typedef {string} MarketSymbol Market symbol (like: 'BTCEUR' or 'KRAKEN:BTCEUR')
+ */
+
+module.exports = {};
